+++ pkgs/dart_mcp_server/test/dart_tooling_mcp_server_test.dart
@@ -6,6 +6,9 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:dart_mcp/server.dart';
+import 'package:dart_mcp_server/src/arg_parser.dart';
+import 'package:dart_mcp_server/src/mixins/analyzer.dart';
+import 'package:dart_mcp_server/src/mixins/dtd.dart';
 import 'package:dart_mcp_server/src/server.dart';
 import 'package:dart_mcp_server/src/utils/analytics.dart';
 import 'package:test/test.dart';
@@ -255,6 +258,28 @@ void main() {
       // Wait for the process to release the file.
       await doWithRetries(() => File(logDescriptor.io.path).delete());
     });
+  });
+
+  test('Tools can be disabled with --exclude-tool', () async {
+    final testHarness = await TestHarness.start(
+      cliArgs: [
+        '--$excludeToolOption',
+        DartAnalyzerSupport.analyzeFilesTool.name,
+        '--$excludeToolOption',
+        DartToolingDaemonSupport.connectTool.name,
+      ],
+    );
+    final connection = testHarness.serverConnectionPair.serverConnection;
+    final tools = (await connection.listTools()).tools;
+    expect(
+      tools,
+      isNot(contains(equals(DartAnalyzerSupport.analyzeFilesTool))),
+    );
+    expect(
+      tools,
+      isNot(contains(equals(DartToolingDaemonSupport.connectTool))),
+    );
+    expect(tools, contains(equals(DartAnalyzerSupport.hoverTool)));
   });
 }